+++ app/src/main/java/nextstep/payments/screens/card/CardCompanyMapper.kt
@@ -1,6 +1,5 @@
 package nextstep.payments.screens.card
 
-import nextstep.payments.CardCompanyState
 import nextstep.payments.domain.CardCompany
 
 fun CardCompanyState.toDomain() = CardCompany(
