+++ app/src/main/java/nextstep/payments/screens/card/new/components/CardCompanyBottomSheetDialog.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import nextstep.payments.CardCompanyState
+import nextstep.payments.screens.card.CardCompanyState
 import nextstep.payments.ui.theme.PaymentsTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
